Extract creation / watching logic to lib to re-use for cron
