+++ bin/create.js
@@ -1,7 +1,7 @@
 var Bluebird = require('bluebird');
 var Bundler = require('webtask-bundle');
 var Chalk = require('chalk');
-var Cli = require('../cli');
+var Cli = require('structured-cli');
 var Concat = require('concat-stream');
 var ConfigFile = require('../lib/config');
 var Fs = Bluebird.promisifyAll(require('fs'));
@@ -10,10 +10,11 @@ var Path = require('path');
 var Sandbox = require('sandboxjs');
 var Superagent = require('superagent');
 var Watcher = require('filewatcher');
+var WebtaskCreator = require('../lib/webtaskCreator');
 var _ = require('lodash');
 
 
-module.exports = Cli.command('create', {
+module.exports = Cli.createCommand('create', {
     description: 'Get information about an existing webtask profile',
     handler: handleCreate,
     options: {
@@ -157,151 +158,25 @@ function handleCreate(args) {
     var config = new ConfigFile();
 
     return config.getProfile(args.profile)
-        .then(createWebtask);
-    
-    
-    function createWebtask(profile) {
-        return args.source === 'url'
-            ?   createSimpleWebtask(profile)
-            :   createLocalWebtask(profile);
-    }
-    
-    function createSimpleWebtask(profile, logger) {
-        if (!logger) logger = _.bindKey(console, 'log');
+        .then(onProfile);
         
-        var codeOrUrl$ = args.capture
-            ?   Sandbox.issueRequest(Superagent.get(args.spec)).get('text')
-            :   args.source === 'stdin'
-                ?   readStdin()
-                :   Bluebird.resolve(args.spec);
-        
-        var webtask$ = codeOrUrl$
-            .then(function (codeOrUrl) {
-                return profile.create(codeOrUrl, {
-                    name: args.name,
-                    merge: args.merge,
-                    parse: args.parse,
-                    secrets: args.secrets,
-                });
-            });
-        
-        return webtask$
-            .then(function (webtask) {
-                logger(webtask.url);
-            });
-        
-        
-        function readStdin() {
-            return new Bluebird(function (resolve, reject) {
-                var concat = Concat({ encoding: 'string' }, resolve);
-                
-                concat.once('error', reject);
-                
-                process.stdin.pipe(concat);
-            });
-        }
-    }
     
-    function createLocalWebtask(profile) {
-        return args.bundle
-            ?   createBundledWebtask(profile)
-            :   createSimpleFileWebtask(profile);
-    }
-    
-    function createBundledWebtask(profile) {
-        return new Bluebird(function (resolve, reject) {
-            var lastGeneration;
-            var logger = args.watch
-                ?   _.bindKey(Logs.createLogStream(profile), 'info')
-                :   _.bindKey(console, 'log');
-            
-            Bundler.bundle({
-                entry: args.spec,
-                loose: args.loose,
-                watch: args.watch,
-            })
-                .subscribe(onGeneration, reject, function () {
-                    resolve();
-                });
-                
-            function onGeneration(build) {
-                lastGeneration = build.generation;
-                
-                if (build.stats.errors.length) {
-                    return _.forEach(build.stats.errors, logger);
-                }
-                
-                var webtask$ = profile.create(build.code, {
-                    name: args.name,
-                    merge: args.merge,
-                    parse: args.parse,
-                    secrets: args.secrets,
-                });
-                
-                webtask$
-                    .then(_.partial(onWebtask, lastGeneration), onWebtaskError);
-            }
-            
-            function onWebtask(generation, webtask) {
-                if (lastGeneration === generation) {
-                    logWebtaskGeneration(logger, generation, webtask);
-                }
-            }
-            
-            function onWebtaskError(err) {
-                console.log('onWebtaskError', err);
-                reject(err);
-            }
+    function onProfile(profile) {
+        var createWebtask = WebtaskCreator(args, {
+            onGeneration: onGeneration,
         });
-    }
-    
-    function createSimpleFileWebtask(profile) {
-        return args.watch
-            ?   createWatchedFileWebtask(profile)
-            :   createSimpleWebtask(profile);
-    }
-    
-    function createWatchedFileWebtask(profile) {
-        return new Bluebird(function (resolve, reject) {
-            
-            var logs = Logs.createLogStream(profile);
-            var watcher = Watcher();
-            var queue = Bluebird.resolve();
-            var generation = 0;
-            
-            watcher.add(args.spec);
-            
-            watcher.on('change', onChange);
-            watcher.on('error', onError);
-            
-            onChange();
-            
-            function onChange() {
-                queue = queue.then(function () {
-                    var webtask$ = createSimpleWebtask(profile, logger);
-                    
-                    return webtask$
-                        .catch(onError);
-    
-                    function logger(url) {
-                        logs.info('Completed build at generation %d: %s', generation++, url);
-                    }
-                });
-            }
-            
-            function onError(err) {
-                watcher.removeAll();
-                
-                reject(err);
-            }
-        });
-    }
-    
-    function logWebtaskGeneration(logger, generation, webtask) {
-        args.watch
-            ?   logger('Completed build at generation %d: %s', generation, webtask.url)
-            :   logger(webtask.url);
+        var log = args.watch
+            ?   _.bindKey(Logs.createLogStream(profile), 'info')
+            :   _.bindKey(console, 'log');
 
+        return createWebtask(profile);
+        
+        
+        function onGeneration(build) {
+            args.watch
+                ?   log({ generation: build.generation, container: build.webtask.container }, 'Webtask created: %s', build.webtask.url)
+                :   log(build.webtask.url);
+        }
     }
 }
