+++ lib/config.js
@@ -1,5 +1,5 @@
 var Bluebird = require('bluebird');
-var Cli = require('../cli');
+var Cli = require('structured-cli');
 var Fs = require('fs');
 var Path = require('path');
 var Sandbox = require('sandboxjs');
