+++ lib/logs.js
@@ -19,6 +19,7 @@ function createLogStream(profile, options) {
     });
     
     logStream.once('error', function (err) {
+        console.dir(err.stack);
         logger.error(err.message);
     });
