+++ lib/webtaskCreator.js
@@ -1,0 +1,162 @@
+var Bluebird = require('bluebird');
+var Bundler = require('webtask-bundle');
+var Cli = require('structured-cli');
+var Concat = require('concat-stream');
+var Sandbox = require('sandboxjs');
+var Superagent = require('superagent');
+var Watcher = require('filewatcher');
+var _ = require('lodash');
+
+
+module.exports = createWebtaskCreator;
+
+
+function createWebtaskCreator(args, options) {
+    options = _.defaults({}, options, {
+        onError: _.noop,
+        onGeneration: _.noop,
+    });
+    
+    return createWebtask;
+    
+    
+    function createWebtask(profile) {
+        return args.source === 'url'
+            ?   createSimpleWebtask(profile)
+            :   createLocalWebtask(profile);
+    }
+    
+    function createSimpleWebtask(profile, generation) {
+        var codeOrUrl$ = args.capture
+            ?   Sandbox.issueRequest(Superagent.get(args.spec)).get('text')
+            :   args.source === 'stdin'
+                ?   readStdin()
+                :   Bluebird.resolve(args.spec);
+        
+        var webtask$ = codeOrUrl$
+            .then(function (codeOrUrl) {
+                return profile.create(codeOrUrl, {
+                    name: args.name,
+                    merge: args.merge,
+                    parse: args.parse,
+                    secrets: args.secrets,
+                });
+            });
+        
+        return webtask$
+            .then(function (webtask) {
+                options.onGeneration({
+                    generation: +generation,
+                    webtask: webtask,
+                });
+            });
+        
+        
+        function readStdin() {
+            return new Bluebird(function (resolve, reject) {
+                if (process.stdin.isTTY) {
+                    return reject(Cli.error.invalid('Code must be piped in when no code or url is specified'));
+                }
+                
+                var concat = Concat({ encoding: 'string' }, resolve);
+                
+                concat.once('error', reject);
+                
+                process.stdin.pipe(concat);
+            });
+        }
+    }
+    
+    function createLocalWebtask(profile) {
+        return args.bundle
+            ?   createBundledWebtask(profile)
+            :   createSimpleFileWebtask(profile);
+    }
+    
+    function createBundledWebtask(profile) {
+        return new Bluebird(function (resolve, reject) {
+            var lastGeneration;
+            
+            Bundler.bundle({
+                entry: args.spec,
+                loose: args.loose,
+                watch: args.watch,
+            })
+                .subscribe(onGeneration, reject, function () {
+                    resolve();
+                });
+                
+            function onGeneration(build) {
+                lastGeneration = build.generation;
+                
+                if (build.stats.errors.length) {
+                    return _.forEach(build.stats.errors, options.onError);
+                }
+                
+                var webtask$ = profile.create(build.code, {
+                    name: args.name,
+                    merge: args.merge,
+                    parse: args.parse,
+                    secrets: args.secrets,
+                });
+                
+                webtask$
+                    .then(_.partial(onWebtask, lastGeneration), _.partial(onWebtaskError, lastGeneration));
+                
+                
+                function onWebtask(generation, webtask) {
+                    if (lastGeneration === generation) {
+                        options.onGeneration({
+                            generation: generation,
+                            webtask: webtask,
+                        });
+                    }
+                }
+                
+                function onWebtaskError(generation, err) {
+                    if (lastGeneration === generation) {
+                        console.log('onWebtaskError', err.stack);
+                        reject(err);
+                    }
+                }
+            }
+        });
+    }
+    
+    function createSimpleFileWebtask(profile) {
+        return args.watch
+            ?   createWatchedFileWebtask(profile)
+            :   createSimpleWebtask(profile);
+    }
+    
+    function createWatchedFileWebtask(profile) {
+        return new Bluebird(function (resolve, reject) {
+            
+            var watcher = Watcher();
+            var queue = Bluebird.resolve();
+            var generation = 0;
+            
+            watcher.add(args.spec);
+            
+            watcher.on('change', onChange);
+            watcher.on('error', onError);
+            
+            onChange();
+            
+            function onChange() {
+                queue = queue.then(function () {
+                    var webtask$ = createSimpleWebtask(profile, ++generation);
+                    
+                    return webtask$
+                        .catch(onError);
+                });
+            }
+            
+            function onError(err) {
+                watcher.removeAll();
+                
+                reject(err);
+            }
+        });
+    }
+}
